[i223] put audio style into message list/composer
